Added flag system
